replace slicer class completely
